+++ app/(layout_HF)/main/_components/organisms/HotRollpeList.tsx
@@ -1,19 +1,15 @@
 "use client";
 import MainRollpeCard from "@/app/_components/molecules/ui/card/main-rollpe-list/MainRollpeCard";
 import styled from "styled-components";
-import { Rollpe } from "@/public/utils/types";
+import { RollpeListProps } from "@/public/utils/types";
 
-// 25.06.20
-// 최근 뜨고 있는 롤페 리스트 받아서 출력만 하도록 리팩토링 완료
-// by HarenKei
-
-const HotRollpeList: React.FC<{ hotRollpeList: Rollpe[] }> = ({
+const HotRollpeList: React.FC<{ hotRollpeList: RollpeListProps }> = ({
   hotRollpeList,
 }) => {
   return (
     <ListContainer>
       {hotRollpeList &&
-        hotRollpeList.map((rollpe) => (
+        hotRollpeList.results.map((rollpe) => (
           <MainRollpeCard
             key={rollpe.id}
             receivingDate={rollpe.receive.receivingDate}
