+++ app/(layout_HF)/mypage/invited-rollpe/page.tsx
@@ -1,12 +1,9 @@
-"use client";
-import { COLORS } from "@/public/styles/colors";
-import styled from "styled-components";
-import { RollpeListProps } from "@/public/utils/types";
 import { RollpeList } from "@/app/_components/molecules/list/RollpeList";
-import { useTransition, useState, useEffect } from "react";
-import { Rollpe } from "@/public/utils/types";
-import { useRollpeList } from "@/public/lib/hooks/fetching/rollpe/useRollpeList";
 import localFont from "next/font/local";
+import {
+  MyRollpeWrapper,
+  MyRollpeContainer,
+} from "./_components/styles/InvitedRollpeStyles";
 
 const hakgyoansim = localFont({
   src: "../../../../public/fonts/HakgyoansimR.woff2",
@@ -26,28 +23,5 @@ const InvitedRollpePage: React.FC = () => {
     </MyRollpeWrapper>
   );
 };
-
-const MyRollpeWrapper = styled.main`
-  padding: 5rem 1.25rem;
-  width: calc(100% - 2.5rem);
-`;
-
-const MyRollpeContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 2rem;
-
-  width: 100%;
-
-  & > .title-wrapper > h1 {
-    color: ${COLORS.ROLLPE_SECONDARY};
-    text-align: center;
-    font-size: 2rem;
-    font-style: normal;
-    font-weight: 400;
-    line-height: normal;
-  }
-`;
 
 export default InvitedRollpePage;
